+++ e2e/helpers.ts
@@ -1,0 +1,13 @@
+import type { Page } from '@playwright/test';
+
+export async function resetAppData(page: Page) {
+  await page.goto('/');
+  await page.evaluate(async () => {
+    await new Promise<void>((resolve, reject) => {
+      const request = indexedDB.deleteDatabase('keyval-store');
+      request.onsuccess = () => resolve();
+      request.onerror = () => reject(request.error);
+      request.onblocked = () => resolve();
+    });
+  });
+}
