+++ e2e/workouts.spec.ts
@@ -1,16 +1,5 @@
 import { expect, test, type Page } from '@playwright/test';
-
-async function resetAppData(page: Page) {
-  await page.goto('/');
-  await page.evaluate(async () => {
-    await new Promise<void>((resolve, reject) => {
-      const request = indexedDB.deleteDatabase('keyval-store');
-      request.onsuccess = () => resolve();
-      request.onerror = () => reject(request.error);
-      request.onblocked = () => resolve();
-    });
-  });
-}
+import { resetAppData } from './helpers';
 
 async function openWorkoutsPage(page: Page) {
   await resetAppData(page);
